+++ src/components/WorkList/index.js
@@ -1,0 +1,2 @@
+import WorkList from './WorkList';
+export default WorkList;
