+++ src/components/style.scss
@@ -703,61 +703,6 @@ body.portal-open {
   height: 100%;
 }
 
-.tabs {
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr;
-  max-width: 1280px;
-  margin: 2rem auto 3.5rem;
-  padding-left: 0;
-  list-style: none;
-}
-
-.tab {
-  padding: 0.5rem;
-  text-align: center;
-  transition: opacity 250ms ease-in;
-
-  &__content {
-    display: block;
-    opacity: 0.33;
-    overflow: hidden;
-  }
-
-  &:hover .tab__content {
-    opacity: 0.66;
-  }
-
-  &:focus {
-    outline: none;
-
-    .tab__icon svg {
-      filter: drop-shadow(0 0 4px #fff);
-    }
-  }
-
-  &.active {
-    .tab__content {
-      opacity: 1;
-    }
-  }
-
-  &__icon {
-    display: inline-block;
-    width: 3.75rem;
-    max-width: 120px;
-  }
-
-  &__title {
-    display: block;
-  }
-}
-
-.tabPanel {
-  > * + * {
-    margin-top: 2em;
-  }
-}
-
 .icon {
   position: relative;
   display: block;
@@ -772,82 +717,6 @@ body.portal-open {
   svg {
     @extend .absolute-fill;
     fill: currentColor;
-  }
-}
-
-.case {
-  &__left {
-    text-align: center;
-  }
-
-  &__year {
-    margin: 0 auto;
-    word-break: break-all;
-    width: 2rem;
-  }
-
-  &__arrow {
-    width: 0.75em;
-    margin: 1em auto 0;
-
-    .icon {
-      padding-bottom: 671.9%;
-    }
-  }
-
- &__thumb {
-    position: relative;
-    transition: .2s ease;
-
-    &::before {
-      content: '';
-      z-index: 1;
-      opacity: 0;
-      @extend .absolute-fill;
-      background-color: rgba(0,0,0,0.8);
-      mix-blend-mode: multiply;
-      transition: inherit;
-    }
-
-    &:hover::before {
-      opacity: 1;
-    }
-
-    &:hover .case__meta {
-      opacity: 1;
-    }
-  }
-
-  &__meta {
-    z-index: 1;
-    opacity: 0;
-    @extend .absolute-fill;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-direction: column;
-    padding: 1rem;
-    transition: inherit;
-  }
-
-  &__icon {
-    width: 7.125rem;
-  }
-
-  &__title {
-    @extend .heading;
-    @extend .h3;
-    margin-top: 1em;
-    text-align: center;
-  }
-
-  &__divider {
-    width: 100%;
-    padding: 4em 17.175%;
-
-    .icon {
-      padding-bottom: 18.125%;
-    }
   }
 }
 
@@ -910,11 +779,17 @@ body.portal-open {
     }
   }
 
+  &__meta {
+    > * + * {
+      margin-top: 1rem;
+    }
+  }
 
   &__description {
     line-height: 1.33;
     font-size: 18px;
     font-style: italic;
+    white-space: pre-line;
 
     @media screen and (min-width: #{$min-width-md}) {
       font-size: 0.75rem;
@@ -922,17 +797,22 @@ body.portal-open {
   }
 
   &__pagination {
+    width: 100%;
     margin-top: 1rem;
-  }
 
-  .icon-button {
-    margin-top: 1rem;
+    @media screen and (min-width: 501px) {
+      width: 15rem;
+      max-width: 100%;
+      margin-left: auto;
+      margin-right: auto;
+    }
   }
 
   .wave {
     width: 10rem;
     max-width: 100%;
-    margin: 1rem auto;
+    margin-left: auto;
+    margin-right: auto;
   }
 
   .eye {
