+++ src/pages/work/case-25-12.md
@@ -2,12 +2,11 @@
 templateKey: work-item
 title: Η περίπτωση 25-12
 titleEN: Case 25-12
-date: 2017-12-31T22:53:20.350Z
+date: 2018-01-01T22:53:20.350Z
 type: painting
 image: /img/stelios-faitakis-greek-artist-case2512-2018.jpg
 description: |+
   60 X 60
-
 descriptionEN: |+
   60 X 60
 ---
