+++ static/admin/config.yml
@@ -69,10 +69,10 @@ collections:
           - {label: "Title", name: "title", widget: "string"}
           - {label: "Title EN", name: "titleEN", widget: "string"}
           - {label: "Description", name: "description", widget: "text"}
-          - {label: "Description EN", name: "description", widget: "text"}
+          - {label: "Description EN", name: "descriptionEN", widget: "text"}
           - {label: "Content", name: "bio", widget: "object", fields: [{label: "Paragraph 1", name: "paragraph1", widget: "object", fields: [{label: "EL", name: "el", widget: "string"}, {label: "EN", name: "en", widget: "string"}]}, {label: "Paragraph 2", name: "paragraph2", widget: "object", fields: [{label: "EL", name: "el", widget: "string"}, {label: "EN", name: "en", widget: "string"}]}, {label: "Paragraph 3", name: "paragraph3", widget: "object", fields: [{label: "EL", name: "el", widget: "string"}, {label: "EN", name: "en", widget: "string"}]}]}
           - {label: "Image", name: "image", widget: "image"}
-          - {label: "Timeline Title", name: "timelineTitle", widget: "object", fields: [{label: 'EL', name: 'el', widget: 'string'}, {label: 'EN', name: 'en', widget: 'string'}]}
+          - {label: "Timeline Title", name: "timelineTitle", widget: "object", fields: [{label: 'EL', name: 'elTimelineTitle', widget: 'string'}, {label: 'EN', name: 'enTimelineTitle', widget: 'string'}]}
           - {label: "Timeline Events", name: "timelineEvents", widget: list, fields: [{ label: Year, name: year, widget: string }, { label: Title, name: title, widget: string }, { label: Description, name: description, widget: string }]}
       - file: "src/pages/work/index.md"
         label: "Work"
@@ -111,4 +111,4 @@ collections:
           - {label: "Title EN", name: "titleEN", widget: "string"}
           - {label: "Description", name: "description", widget: "text"}
           - {label: "Description EN", name: "descriptionEN", widget: "text"}
-          - {label: "Content", name: "terms", widget: "object", fields: [{label: "EL", name: "el", widget: "markdown"}, {label: "EN", name: "en", widget: "markdown"}]}
+          - {label: "Content", name: "terms", widget: "object", fields: [{label: "EL", name: "elTerms", widget: "markdown"}, {label: "EN", name: "enTerms", widget: "markdown"}]}
